multitenant: relocate `tenantcapabilitiespb` directory
This commit relocates the `tenantcapabilitiespb` directory to the same
level as `tenantcapabilities`.  This change improves consistency with other
similar directories (e.g., `mtinfopb`) and shortens the long package path
(`pkg/multitenant/tenantcapabilities`).
