+++ pkg/multitenant/tenantcapabilities/capabilities.go
@@ -6,7 +6,7 @@
 package tenantcapabilities
 
 import (
-	"github.com/cockroachdb/cockroach/pkg/multitenant/tenantcapabilities/tenantcapabilitiespb"
+	"github.com/cockroachdb/cockroach/pkg/multitenant/tenantcapabilitiespb"
 	"github.com/cockroachdb/cockroach/pkg/spanconfig/spanconfigbounds"
 	"github.com/cockroachdb/errors"
 	"github.com/cockroachdb/redact"
