+++ pkg/multitenant/tenantcapabilities/capability.go
@@ -9,7 +9,7 @@ package tenantcapabilities
 import (
 	"fmt"
 
-	"github.com/cockroachdb/cockroach/pkg/multitenant/tenantcapabilities/tenantcapabilitiespb"
+	"github.com/cockroachdb/cockroach/pkg/multitenant/tenantcapabilitiespb"
 	"github.com/cockroachdb/cockroach/pkg/spanconfig/spanconfigbounds"
 	"github.com/cockroachdb/redact"
 	"github.com/cockroachdb/redact/interfaces"
