+++ pkg/multitenant/tenantcapabilities/tenantcapabilitieswatcher/BUILD.bazel
@@ -18,7 +18,7 @@ go_library(
         "//pkg/kv/kvpb",
         "//pkg/multitenant/mtinfo",
         "//pkg/multitenant/tenantcapabilities",
-        "//pkg/multitenant/tenantcapabilities/tenantcapabilitiespb",
+        "//pkg/multitenant/tenantcapabilitiespb",
         "//pkg/roachpb",
         "//pkg/settings/cluster",
         "//pkg/sql/catalog",
@@ -52,7 +52,7 @@ go_test(
         "//pkg/kv/kvclient/rangefeed/rangefeedcache",
         "//pkg/multitenant/mtinfopb",
         "//pkg/multitenant/tenantcapabilities",
-        "//pkg/multitenant/tenantcapabilities/tenantcapabilitiespb",
+        "//pkg/multitenant/tenantcapabilitiespb",
         "//pkg/multitenant/tenantcapabilities/tenantcapabilitiestestutils",
         "//pkg/roachpb",
         "//pkg/security/securityassets",
