+++ pkg/multitenant/tenantcapabilities/values.go
@@ -6,7 +6,7 @@
 package tenantcapabilities
 
 import (
-	"github.com/cockroachdb/cockroach/pkg/multitenant/tenantcapabilities/tenantcapabilitiespb"
+	"github.com/cockroachdb/cockroach/pkg/multitenant/tenantcapabilitiespb"
 	"github.com/cockroachdb/errors"
 )
