+++ pkg/multitenant/tenantcapabilities/values_test.go
@@ -8,7 +8,7 @@ package tenantcapabilities
 import (
 	"testing"
 
-	"github.com/cockroachdb/cockroach/pkg/multitenant/tenantcapabilities/tenantcapabilitiespb"
+	"github.com/cockroachdb/cockroach/pkg/multitenant/tenantcapabilitiespb"
 	"github.com/cockroachdb/errors"
 	"github.com/stretchr/testify/require"
 )
