+++ pkg/multitenant/tenantcapabilitiespb/BUILD.bazel
@@ -16,7 +16,7 @@ proto_library(
 go_proto_library(
     name = "tenantcapabilitiespb_go_proto",
     compilers = ["//pkg/cmd/protoc-gen-gogoroach:protoc-gen-gogoroach_compiler"],
-    importpath = "github.com/cockroachdb/cockroach/pkg/multitenant/tenantcapabilities/tenantcapabilitiespb",
+    importpath = "github.com/cockroachdb/cockroach/pkg/multitenant/tenantcapabilitiespb",
     proto = ":tenantcapabilitiespb_proto",
     visibility = ["//visibility:public"],
     deps = [
@@ -28,6 +28,6 @@ go_proto_library(
 go_library(
     name = "tenantcapabilitiespb",
     embed = [":tenantcapabilitiespb_go_proto"],
-    importpath = "github.com/cockroachdb/cockroach/pkg/multitenant/tenantcapabilities/tenantcapabilitiespb",
+    importpath = "github.com/cockroachdb/cockroach/pkg/multitenant/tenantcapabilitiespb",
     visibility = ["//visibility:public"],
 )
