+++ pkg/protos.bzl
@@ -18,7 +18,7 @@ SERVER_PROTOS = [
     "//pkg/kv/kvserver/loqrecovery/loqrecoverypb:loqrecoverypb_proto",
     "//pkg/kv/kvserver/readsummary/rspb:rspb_proto",
     "//pkg/multitenant/mtinfopb:mtinfopb_proto",
-    "//pkg/multitenant/tenantcapabilities/tenantcapabilitiespb:tenantcapabilitiespb_proto",
+    "//pkg/multitenant/tenantcapabilitiespb:tenantcapabilitiespb_proto",
     "//pkg/raft/raftpb:raftpb_proto",
     "//pkg/roachpb:roachpb_proto",
     "//pkg/rpc/rpcpb:rpcpb_proto",
@@ -89,7 +89,7 @@ PROTO_FILES = [
     "//pkg/kv/kvserver/loqrecovery/loqrecoverypb:recovery.proto",
     "//pkg/kv/kvserver/readsummary/rspb:summary.proto",
     "//pkg/multitenant/mtinfopb:info.proto",
-    "//pkg/multitenant/tenantcapabilities/tenantcapabilitiespb:capabilities.proto",
+    "//pkg/multitenant/tenantcapabilitiespb:capabilities.proto",
     "//pkg/raft/raftpb:raft.proto",
     "//pkg/roachpb:data.proto",
     "//pkg/roachpb:index_usage_stats.proto",
