+++ pkg/spanconfig/spanconfigbounds/BUILD.bazel
@@ -20,7 +20,7 @@ go_library(
     visibility = ["//visibility:public"],
     deps = [
         "//pkg/config",
-        "//pkg/multitenant/tenantcapabilities/tenantcapabilitiespb",
+        "//pkg/multitenant/tenantcapabilitiespb",
         "//pkg/roachpb",
         "//pkg/util/protoutil",
         "@com_github_cockroachdb_errors//:errors",
@@ -38,7 +38,7 @@ go_test(
     data = glob(["testdata/**"]),
     embed = [":spanconfigbounds"],
     deps = [
-        "//pkg/multitenant/tenantcapabilities/tenantcapabilitiespb",
+        "//pkg/multitenant/tenantcapabilitiespb",
         "//pkg/roachpb",
         "//pkg/testutils/datapathutils",
         "//pkg/util/protoutil",
