+++ pkg/spanconfig/spanconfigbounds/bounds.go
@@ -8,7 +8,7 @@ package spanconfigbounds
 import (
 	"fmt"
 
-	"github.com/cockroachdb/cockroach/pkg/multitenant/tenantcapabilities/tenantcapabilitiespb"
+	"github.com/cockroachdb/cockroach/pkg/multitenant/tenantcapabilitiespb"
 	"github.com/cockroachdb/cockroach/pkg/roachpb"
 	"github.com/cockroachdb/redact"
 )
