+++ pkg/sql/BUILD.bazel
@@ -347,7 +347,7 @@ go_library(
         "//pkg/multitenant/mtinfopb",
         "//pkg/multitenant/multitenantcpu",
         "//pkg/multitenant/tenantcapabilities",
-        "//pkg/multitenant/tenantcapabilities/tenantcapabilitiespb",
+        "//pkg/multitenant/tenantcapabilitiespb",
         "//pkg/repstream",
         "//pkg/repstream/streampb",
         "//pkg/roachpb",
